requirements.txt
certifi>=2018
slackclient>=2.0.0
